+++ level_compiler/src/trigger_parse_state.rs
@@ -4,8 +4,8 @@ pub struct TriggerParseState {
 	pub kind: Option<TriggerKindSource>,
 	pub top: i32,
 	pub left: i32,
-	pub width: i32,
-	pub height: i32,
+	pub width: f32,
+	pub height: f32,
 	pub icon_id: i32,
 }
 
@@ -15,8 +15,8 @@ impl TriggerParseState {
 			kind: None,
 			top: 0,
 			left: 0,
-			width: 1,
-			height: 1,
+			width: 1.0,
+			height: 1.0,
 			icon_id: 0,
 		};
 	}
@@ -25,8 +25,8 @@ impl TriggerParseState {
 		self.kind = None;
 		self.top = 0;
 		self.left = 0;
-		self.width = 1;
-		self.height = 1;
+		self.width = 1.0;
+		self.height = 1.0;
 		self.icon_id = 0;
 	}
